+++ app/controllers/Panel/Keanggotaan/PanelKeanggotaanHpController.php
@@ -85,6 +85,8 @@ class PanelKeanggotaanHpController extends BaseController {
 	 */
 	public function destroy($anggota, $hp)
 	{
+		if(! Input::get('safe-action')) return Redirect::back();
+
 		$hp->delete();
 		return Redirect::action('panel.keanggotaan.hp.index', $anggota->id_anggota)->with('success', 'No. Handphone berhasil dihapus!');
 	}
